requirements.txt

asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
httpretty==0.9.5
idna==2.7
mock==2.0.0
ndg-httpsclient==0.5.1
nose==1.3.7
pbr==5.0.0
pyasn1==0.4.4
pycparser==2.19
pyOpenSSL==18.0.0
requests==2.20.0
six==1.11.0
tornado==5.1.1
urllib3==1.24
aiohttp==3.4.4; python_version >= '3.5'
async-timeout==3.0.1; python_version >= '3.5'
attrs==18.2.0; python_version >= '3.5'
idna-ssl==1.1.0; python_version >= '3.5' and python_version < '3.7'
multidict==4.4.2; python_version >= '3.5'
yarl==1.2.6; python_version >= '3.5'
pynacl==1.3.0
